tengo miedo2

+++ TipicoSaludable2/app/src/main/java/com/sanchez/tipicosaludable/ScrollingDetalle.java
@@ -173,11 +173,10 @@ public class ScrollingDetalle extends AppCompatActivity {
                     if (cantidaddelalimento>=20){
                         edtxcantidad.setError("No puedes comer tanto");
                         edtxcantidad.setText("");
-
                     }
                     else {
 
-                        Calorias_consumidas = cantidaddelalimento*(consumo+ Integer.parseInt(informacion.getText().toString()));
+                        Calorias_consumidas = cantidaddelalimento*(consumo+Double.parseDouble(informacion.getText().toString()));
                         Toast.makeText(ScrollingDetalle.this, ""+Calorias_consumidas, Toast.LENGTH_SHORT).show();
                         Calorias_consumidas = cantidaddelalimento*(consumo+ Double.parseDouble(informacion.getText().toString()));
                         //Toast.makeText(ActividadDetalle.this, ""+Calorias_consumidas, Toast.LENGTH_SHORT).show();
